+++ sdk/search/azure-search-documents/azure/search/documents/indexes/_search_index_client.py
@@ -3,7 +3,7 @@
 # Licensed under the MIT License. See License.txt in the project root for
 # license information.
 # --------------------------------------------------------------------------
-from typing import Any, Dict, List, Union, Optional
+from typing import Any, List, Union, Optional, MutableMapping, cast
 
 from azure.core import MatchConditions
 from azure.core.credentials import AzureKeyCredential, TokenCredential
@@ -20,7 +20,13 @@ from .._headers_mixin import HeadersMixin
 from .._utils import get_authentication_policy
 from .._version import SDK_MONIKER
 from .._search_client import SearchClient
-from .models import SearchIndex, SynonymMap, SearchAlias, AnalyzeTextOptions, AnalyzeResult
+from .models import (
+    SearchIndex,
+    SynonymMap,
+    SearchAlias,
+    AnalyzeTextOptions,
+    AnalyzeResult,
+)
 
 
 class SearchIndexClient(HeadersMixin):  # pylint:disable=too-many-public-methods
@@ -37,6 +43,7 @@ class SearchIndexClient(HeadersMixin):  # pylint:disable=too-many-public-methods
     """
 
     _ODATA_ACCEPT: str = "application/json;odata.metadata=minimal"
+    _client: _SearchServiceClient
 
     def __init__(self, endpoint: str, credential: Union[AzureKeyCredential, TokenCredential], **kwargs: Any) -> None:
         self._api_version = kwargs.pop("api_version", DEFAULT_VERSION)
@@ -45,13 +52,13 @@ class SearchIndexClient(HeadersMixin):  # pylint:disable=too-many-public-methods
         audience = kwargs.pop("audience", None)
         if isinstance(credential, AzureKeyCredential):
             self._aad = False
-            self._client: _SearchServiceClient = _SearchServiceClient(
+            self._client = _SearchServiceClient(
                 endpoint=endpoint, sdk_moniker=SDK_MONIKER, api_version=self._api_version, **kwargs
             )
         else:
             self._aad = True
             authentication_policy = get_authentication_policy(credential, audience=audience)
-            self._client: _SearchServiceClient = _SearchServiceClient(
+            self._client = _SearchServiceClient(
                 endpoint=endpoint,
                 authentication_policy=authentication_policy,
                 sdk_moniker=SDK_MONIKER,
@@ -98,7 +105,8 @@ class SearchIndexClient(HeadersMixin):  # pylint:disable=too-many-public-methods
         if select:
             kwargs["select"] = ",".join(select)
         # pylint:disable=protected-access
-        return self._client.indexes.list(cls=lambda objs: [SearchIndex._from_generated(x) for x in objs], **kwargs)
+        indexes = self._client.indexes.list(cls=lambda objs: [SearchIndex._from_generated(x) for x in objs], **kwargs)
+        return cast(ItemPaged[SearchIndex], indexes)
 
     @distributed_trace
     def list_index_names(self, **kwargs: Any) -> ItemPaged[str]:
@@ -111,7 +119,8 @@ class SearchIndexClient(HeadersMixin):  # pylint:disable=too-many-public-methods
         """
         kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
 
-        return self._client.indexes.list(cls=lambda objs: [x.name for x in objs], **kwargs)
+        names = self._client.indexes.list(cls=lambda objs: [x.name for x in objs], **kwargs)
+        return cast(ItemPaged[str], names)
 
     @distributed_trace
     def get_index(self, name: str, **kwargs: Any) -> SearchIndex:
@@ -137,7 +146,7 @@ class SearchIndexClient(HeadersMixin):  # pylint:disable=too-many-public-methods
         return SearchIndex._from_generated(result)  # pylint:disable=protected-access
 
     @distributed_trace
-    def get_index_statistics(self, index_name: str, **kwargs: Any) -> Dict:
+    def get_index_statistics(self, index_name: str, **kwargs: Any) -> MutableMapping[str, Any]:
         """Returns statistics for the given index, including a document count
         and storage usage.
 
@@ -182,7 +191,7 @@ class SearchIndexClient(HeadersMixin):  # pylint:disable=too-many-public-methods
         error_map, access_condition = get_access_conditions(index, match_condition)
         kwargs.update(access_condition)
         try:
-            index_name = index.name
+            index_name = index.name  # type: ignore
         except AttributeError:
             index_name = index
         self._client.indexes.delete(index_name=index_name, error_map=error_map, **kwargs)
@@ -318,6 +327,7 @@ class SearchIndexClient(HeadersMixin):  # pylint:disable=too-many-public-methods
         if select:
             kwargs["select"] = ",".join(select)
         result = self._client.synonym_maps.list(**kwargs)
+        assert result.synonym_maps is not None  # Hint for mypy
         # pylint:disable=protected-access
         return [SynonymMap._from_generated(x) for x in result.synonym_maps]
 
@@ -332,6 +342,7 @@ class SearchIndexClient(HeadersMixin):  # pylint:disable=too-many-public-methods
         """
         kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
         result = self._client.synonym_maps.list(**kwargs)
+        assert result.synonym_maps is not None  # Hint for mypy
         return [x.name for x in result.synonym_maps]
 
     @distributed_trace
@@ -391,7 +402,7 @@ class SearchIndexClient(HeadersMixin):  # pylint:disable=too-many-public-methods
         error_map, access_condition = get_access_conditions(synonym_map, match_condition)
         kwargs.update(access_condition)
         try:
-            name = synonym_map.name
+            name = synonym_map.name  # type: ignore
         except AttributeError:
             name = synonym_map
         self._client.synonym_maps.delete(synonym_map_name=name, error_map=error_map, **kwargs)
@@ -453,7 +464,7 @@ class SearchIndexClient(HeadersMixin):  # pylint:disable=too-many-public-methods
         return SynonymMap._from_generated(result)  # pylint:disable=protected-access
 
     @distributed_trace
-    def get_service_statistics(self, **kwargs: Any) -> Dict:
+    def get_service_statistics(self, **kwargs: Any) -> MutableMapping[str, Any]:
         """Get service level statistics for a search service.
 
         :return: Service statistics result.
@@ -479,7 +490,7 @@ class SearchIndexClient(HeadersMixin):  # pylint:disable=too-many-public-methods
         if select:
             kwargs["select"] = ",".join(select)
         # pylint:disable=protected-access
-        return self._client.aliases.list(**kwargs)
+        return cast(ItemPaged[SearchAlias], self._client.aliases.list(**kwargs))
 
     @distributed_trace
     def list_alias_names(self, **kwargs: Any) -> ItemPaged[str]:
@@ -492,7 +503,8 @@ class SearchIndexClient(HeadersMixin):  # pylint:disable=too-many-public-methods
         """
         kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
 
-        return self._client.aliases.list(cls=lambda objs: [x.name for x in objs], **kwargs)
+        names = self._client.aliases.list(cls=lambda objs: [x.name for x in objs], **kwargs)
+        return cast(ItemPaged[str], names)
 
     @distributed_trace
     def get_alias(self, name: str, **kwargs: Any) -> SearchAlias:
@@ -537,7 +549,7 @@ class SearchIndexClient(HeadersMixin):  # pylint:disable=too-many-public-methods
         error_map, access_condition = get_access_conditions(alias, match_condition)
         kwargs.update(access_condition)
         try:
-            alias_name = alias.name
+            alias_name = alias.name  # type: ignore
         except AttributeError:
             alias_name = alias
         self._client.aliases.delete(alias_name=alias_name, error_map=error_map, **kwargs)